Fixing type conversion warning

+++ src/openrct2-ui/windows/Scenery.cpp
@@ -1232,7 +1232,7 @@ private:
 
     int32_t GetTabRowCount()
     {
-        return std::max<int32_t>((_tabEntries.size() + MaxTabsPerRow - 1) / MaxTabsPerRow, 0);
+        return std::max<int32_t>((static_cast<int32_t>(_tabEntries.size()) + MaxTabsPerRow - 1) / MaxTabsPerRow, 0);
     }
 
     int32_t GetMaxTabCountInARow()
